Prevent removing i3/config so chromebook retains its link to alternative config file.

+++ Makefile
@@ -27,7 +27,7 @@ relink:
 	$(MAKE) ~/.oh-my-zsh
 	$(MAKE) ~/.config/direnv
 	mkdir -p ~/.config/i3
-	rm -f ~/.zshrc ~/.tmux.conf ~/.inputrc ~/.gitconfig ~/.gitignore_global ~/.rspec ~/.spacemacs ~/.config/i3/config
+	rm -f ~/.zshrc ~/.tmux.conf ~/.inputrc ~/.gitconfig ~/.gitignore_global ~/.rspec ~/.spacemacs
 	ln -s ${HOME}/.config/dotfiles/zsh/zshenv ~/.zshenv
 	ln -s ${HOME}/.config/dotfiles/zsh/zshrc ~/.zshrc
 	ln -s ${HOME}/.config/dotfiles/etc/tmux.conf ~/.tmux.conf
@@ -35,7 +35,7 @@ relink:
 	ln -s ${HOME}/.config/dotfiles/etc/gitconfig ~/.gitconfig
 	ln -s ${HOME}/.config/dotfiles/etc/rspec ~/.rspec
 	ln -s ${HOME}/.config/dotfiles/spacemacs/spacemacs ~/.spacemacs
-	ln -s ${HOME}/.config/dotfiles/i3/config ~/.config/i3/config
+	[ ! -f ~/.config/i3/config ] && ln -s ${HOME}/.config/dotfiles/i3/config ~/.config/i3/config
 	# ln -s ${HOME}/.config/dotfiles/vim/vimrc.mine ~/.vim/vimrc.mine
 	# .gitconfig points directly to the global ignore. don't need it in home
 	# directory.  ln -s ${HOME}/.config/dotfiles/etc/gitignore_global ~/.gitignore_global
